Add AuthRequest message type
There is a fairly substantial amount of extra information that is
helpful to know when the system is authenticating that involves
knowing what client is making the request and what service they are
making the request to.  This commit adds an encapsulating message type
that holds this information and the entity on whose behalf the request
is being made.

Additionaly this commit heavily comments several functions to clarify
intent and design of these functions.