Refactor to using functions on a type.
This commit refactors to using functions on the EMDataStore type.  In
principle the logic and flow remains the same but this removes the
package scoped global variables and permits the entity_manager to now
be built as an interface which will improve the capabilities to mock
the functions for testing higher level systems.  This will also pave
the way for pluggable database systems by using interfaces elsewhere
which will make it more clear to use a persistence layer interface
here.