Refactor of the client interface
Rather than a number of standalone functions the client interface is
now based around functions bound to a client struct. This allows for a
much cleaner interface by setting the serverAddr, serverPort,
clientID, and serviceID on instantiation as well as possibilities such
as maintaining a slice of netAuthClient objects and ranging over them
for spiky services to balance connections over a connection pool.