Remove the Entity Manager cache.
Previously the entity_manager had maintained an internal cache of data
for the entities and the groups.  This leads to a lot of
complications.  Most importantly it means that the storage layer can't
invalidate the cache since it has no handle up into the
entity_manager.  This commit removes the cache from the entity_manager
and refactors so that all data requests hit the storage layer.
Storage systems can implement caches that invalidate as needed based
on the constraints of cluster synchronization and similar concerns.