Release v0.1.0
This is a momentous release as the moving of the decimal one to the
right signifies that NetAuth is now out of Alpha and is usable as Beta
quality software.  It is also now in use at a handful of production
environments, and so it makes sense to finally be up into at least the
beta range.

This release also is the first to have binary releases, which includes
all the licenses for all other modules that came with the software.
These licenses are collected via vendor-licenses, which is MIT
licensed and used from Noel Cower's xq-api with permission.