internal/db/protodb: Fix disk access race
This commit fixes a data race pointed out by @m-wynn which could cause
persisted on disk corruption in the event that two carefully timed
events tried to write to the same object on disk.  The solution was to
ensure that all writes happen atomically, and to that end renameio is
imported to handle the disk access.