README: remove production warning.
Though the warning still stands that you shouldn't run things in prod
you don't understand, and you really shouldn't trust your
authentication to things you don't understand, NetAuth has now been
running for Void as an internet facing service for over a year.  It's
also been securing the site-to-site communications of 2 private
networks the author is aware of for more than 6 months, and so one can
be reasonably certain it is not equivalent to a wide open door.

I'm also stripping the warning off because when I wrote it, I was of
the opininion that you had to be somehow special to write security
software.  I'm not sure what I thought this was, whether it was a
special degree or some sword from a lady in a lake or something else
entirely, but I thought I didn't have it.  As I've gone on and worked
more on NetAuth over the last 2 years, I've learned that no-one has
this.  We're all just as clueless and just as likely to make mistakes
as we always have been because the average open source developer is
human (I suspect this to be true of closed source developers, but have
insufficient data at this time).

The code has tests, the tests pass, and the rule of simplicity is
aggressively applied near authentication mechanisms.  If these
approaches can't provide some sense that the code is working
correctly, then it is my opinion that nothing can.  Feel free to run
NetAuth in production, secure your startup with it, and move on with
life.  Of course the terms of the LICENSE still apply, so don't come
to me if it somehow eats your dog or something like that.