internal/rpc2: Implement SystemCapabilities
This commit was supposed to be simple.  It was going to just be
setting capabilities, how hard could this be?  As it turns out really
hard, since parts of the original tree specification take things as
strings that should not be taken as strings.  This lead to
un-detectable errors that would silently be coerced to valid zero
values.  The problem is that one of the valid zero values for a
capability is GLOBAL_ROOT.  So an uninitialized request might try to
grant the GLOBAL_ROOT capability unexpectedly.  This would still
require the caller to have root already, but is certainly unexpected.