internal/db: Refactor to K/V storage
This commit refactors the database package to use key/value storage.
It also includes the various changes needed to keep the various other
packages and tests building after the change.

This change means that supporting new storage systems is much easier,
and that its not nearly as complicated to maintain existing storage
engines.  Most importantly, there are not large amounts of copy/paste
code for dealing with disk IO.