pkg/token/jwt: Refactor JWT provider
This commit does significant refactoring on the JWT-RSA provider.
Among other things it moves the key generation to nsutil out of the
main binary, it clean up a lot of the tests and other code internally,
and it uses KeyProviders to enable keys to be loaded from remote
sources in the future.