nsutil: enable non-interactive bootstrap
We use environment variables so that we discourage as much as possible
anybody leaving these set all the time. These credentials could be used
to take control of any instance, so they should be restricted.

We also allow only bootstrapping this way once, so another provider or
setup could establish a permanent password later and not be wiped out by
a re-auto-bootstrap.

Very useful for test clusters, or automating the setup of real clusters.